+++ kinto/core/errors.py
@@ -1,4 +1,5 @@
 import colander
+import logging
 from pyramid import httpexceptions
 from enum import Enum
 
@@ -193,3 +194,20 @@ def send_alert(request, message=None, url=None, code='soft-eol'):
         'message': message,
         'url': url
     })
+
+
+def request_GET(request):
+    """Catches a UnicodeDecode error in request.GET in case a wrong request was received.
+    Fixing a webob long term issue: https://github.com/Pylons/webob/issues/161
+    """
+    try:
+        return request.GET
+    except UnicodeDecodeError as e:
+        querystring = request.environ.get('QUERY_STRING', '')
+        logger = logging.getLogger(__name__)
+        logger.warn('Error decoding QUERY_STRING: %s' % request.environ)
+        raise http_error(
+            httpexceptions.HTTPBadRequest(),
+            errno=ERRORS.INVALID_PARAMETERS,
+            message="A request with an incorrect encoding in the querystring was"
+            "received. Please make sure your requests are encoded in UTF-8: %s" % querystring)