+++ CHANGELOG.rst
@@ -71,6 +71,7 @@ Protocol is now at version **1.16**. See `API changelog`_.
 - Removed Structlog binding and bottlenecks (fixes #603)
 - Fixed Swagger output with subpath and regex in pyramid routes (fixes #1180)
 - Fixed Postgresql errors when specifying empty values in querystring numeric filters. (fixes #1194)
+- Return a 400 Bad Request instead of crashing when the querystring contains bad characters. (fixes #1195)
 
 **Internal changes**